fix: really random sampling in sql

+++ prompt_enhancing/src/archaeo_super_prompt/dataset/load.py
@@ -8,7 +8,7 @@ from ..cache import memory
 
 
 @memory.cache
-def _init_with_cache(size: int, seed: int):
+def _init_with_cache(size: int, seed: float):
     intervention_data, findings = get_entries(size, seed)
     files = pd.concat(
         [
@@ -26,7 +26,7 @@ def _init_with_cache(size: int, seed: int):
 
 
 class MagohDataset:
-    def __init__(self, size: int, seed: int):
+    def __init__(self, size: int, seed: float):
         """Fetch a maximum of `size` samples from the Magoh training database"""
         self._intervention_data, self._findings, self._files = _init_with_cache(
             size, seed
